+++ src/compiler/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod load;
 pub(crate) mod syntax;
